# src/main/java/org/actics/customer/profiling/api/CustomersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.11.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.actics.customer.profiling.api;

import org.actics.customer.profiling.model.CustomerProfile;
import org.actics.customer.profiling.model.CustomerProfileResponse;
import org.springframework.format.annotation.DateTimeFormat;
import java.time.LocalDate;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;
import java.util.List;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-01-21T22:01:26.297063+01:00[Europe/Zurich]", comments = "Generator version: 7.11.0")
@Validated
@Tag(name = "customers", description = "the customers API")
public interface CustomersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /customers/{customerId} : Retrieve a customer profile
     * Get the details of a customer profile by ID.
     *
     * @param customerId The ID of the customer (required)
     * @return Customer profile details (status code 200)
     *         or Customer not found (status code 404)
     */
    @Operation(
        operationId = "customersCustomerIdGet",
        summary = "Retrieve a customer profile",
        description = "Get the details of a customer profile by ID.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Customer profile details", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CustomerProfileResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Customer not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/customers/{customerId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<CustomerProfileResponse> customersCustomerIdGet(
        @Parameter(name = "customerId", description = "The ID of the customer", required = true, in = ParameterIn.PATH) @PathVariable("customerId") String customerId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"profile\" : { \"firstName\" : \"John\", \"lastName\" : \"Doe\", \"birthdate\" : \"1980-01-01\", \"economicCircumstances\" : { \"annualIncome\" : 100000, \"netWorth\" : 500000, \"liabilities\" : 100000 }, \"investmentExperience\" : \"Moderate\", \"riskTolerance\" : \"Medium\", \"investmentObjectives\" : \"Wealth accumulation\", \"contactDetails\" : { \"phone\" : \"+41791234567\", \"email\" : \"john.doe@example.com\" } }, \"id\" : \"c123456\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /customers : Retrieve all customer profiles
     * Get the details of all customer profiles without filters.
     *
     * @return List of customer profiles (status code 200)
     */
    @Operation(
        operationId = "customersGet",
        summary = "Retrieve all customer profiles",
        description = "Get the details of all customer profiles without filters.",
        responses = {
            @ApiResponse(responseCode = "200", description = "List of customer profiles", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = CustomerProfileResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/customers",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<CustomerProfileResponse>> customersGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"profile\" : { \"firstName\" : \"John\", \"lastName\" : \"Doe\", \"birthdate\" : \"1980-01-01\", \"economicCircumstances\" : { \"annualIncome\" : 100000, \"netWorth\" : 500000, \"liabilities\" : 100000 }, \"investmentExperience\" : \"Moderate\", \"riskTolerance\" : \"Medium\", \"investmentObjectives\" : \"Wealth accumulation\", \"contactDetails\" : { \"phone\" : \"+41791234567\", \"email\" : \"john.doe@example.com\" } }, \"id\" : \"c123456\" }, { \"profile\" : { \"firstName\" : \"John\", \"lastName\" : \"Doe\", \"birthdate\" : \"1980-01-01\", \"economicCircumstances\" : { \"annualIncome\" : 100000, \"netWorth\" : 500000, \"liabilities\" : 100000 }, \"investmentExperience\" : \"Moderate\", \"riskTolerance\" : \"Medium\", \"investmentObjectives\" : \"Wealth accumulation\", \"contactDetails\" : { \"phone\" : \"+41791234567\", \"email\" : \"john.doe@example.com\" } }, \"id\" : \"c123456\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /customers : Create a new customer profile
     * Create a new customer profile with details required for FIDLEG compliance.
     *
     * @param customerProfile  (required)
     * @return Customer profile created successfully (status code 201)
     *         or Invalid input data (status code 400)
     */
    @Operation(
        operationId = "customersPost",
        summary = "Create a new customer profile",
        description = "Create a new customer profile with details required for FIDLEG compliance.",
        responses = {
            @ApiResponse(responseCode = "201", description = "Customer profile created successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CustomerProfileResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input data")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/customers",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<CustomerProfileResponse> customersPost(
        @Parameter(name = "CustomerProfile", description = "", required = true) @Valid @RequestBody CustomerProfile customerProfile
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"profile\" : { \"firstName\" : \"John\", \"lastName\" : \"Doe\", \"birthdate\" : \"1980-01-01\", \"economicCircumstances\" : { \"annualIncome\" : 100000, \"netWorth\" : 500000, \"liabilities\" : 100000 }, \"investmentExperience\" : \"Moderate\", \"riskTolerance\" : \"Medium\", \"investmentObjectives\" : \"Wealth accumulation\", \"contactDetails\" : { \"phone\" : \"+41791234567\", \"email\" : \"john.doe@example.com\" } }, \"id\" : \"c123456\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /customers/search : Search for customer profiles
     * Search customer profiles by attributes such as firstName, lastName, birthdate, riskTolerance, email, or phone.
     *
     * @param firstName Filter by first name (optional)
     * @param lastName Filter by last name (optional)
     * @param birthdate Filter by birthdate (optional)
     * @param riskTolerance Filter by risk tolerance (optional)
     * @param email Filter by email (optional)
     * @param phone Filter by phone number (optional)
     * @return List of matching customer profiles (status code 200)
     *         or Invalid query parameters (status code 400)
     */
    @Operation(
        operationId = "customersSearchGet",
        summary = "Search for customer profiles",
        description = "Search customer profiles by attributes such as firstName, lastName, birthdate, riskTolerance, email, or phone.",
        responses = {
            @ApiResponse(responseCode = "200", description = "List of matching customer profiles", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = CustomerProfileResponse.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid query parameters")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/customers/search",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<CustomerProfileResponse>> customersSearchGet(
        @Parameter(name = "firstName", description = "Filter by first name", in = ParameterIn.QUERY) @Valid @RequestParam(value = "firstName", required = false) String firstName,
        @Parameter(name = "lastName", description = "Filter by last name", in = ParameterIn.QUERY) @Valid @RequestParam(value = "lastName", required = false) String lastName,
        @Parameter(name = "birthdate", description = "Filter by birthdate", in = ParameterIn.QUERY) @Valid @RequestParam(value = "birthdate", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate birthdate,
        @Parameter(name = "riskTolerance", description = "Filter by risk tolerance", in = ParameterIn.QUERY) @Valid @RequestParam(value = "riskTolerance", required = false) String riskTolerance,
        @jakarta.validation.constraints.Email @Parameter(name = "email", description = "Filter by email", in = ParameterIn.QUERY) @Valid @RequestParam(value = "email", required = false) String email,
        @Parameter(name = "phone", description = "Filter by phone number", in = ParameterIn.QUERY) @Valid @RequestParam(value = "phone", required = false) String phone
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"profile\" : { \"firstName\" : \"John\", \"lastName\" : \"Doe\", \"birthdate\" : \"1980-01-01\", \"economicCircumstances\" : { \"annualIncome\" : 100000, \"netWorth\" : 500000, \"liabilities\" : 100000 }, \"investmentExperience\" : \"Moderate\", \"riskTolerance\" : \"Medium\", \"investmentObjectives\" : \"Wealth accumulation\", \"contactDetails\" : { \"phone\" : \"+41791234567\", \"email\" : \"john.doe@example.com\" } }, \"id\" : \"c123456\" }, { \"profile\" : { \"firstName\" : \"John\", \"lastName\" : \"Doe\", \"birthdate\" : \"1980-01-01\", \"economicCircumstances\" : { \"annualIncome\" : 100000, \"netWorth\" : 500000, \"liabilities\" : 100000 }, \"investmentExperience\" : \"Moderate\", \"riskTolerance\" : \"Medium\", \"investmentObjectives\" : \"Wealth accumulation\", \"contactDetails\" : { \"phone\" : \"+41791234567\", \"email\" : \"john.doe@example.com\" } }, \"id\" : \"c123456\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}