# src/main/java/org/actics/customer/profiling/model/CustomerProfileResponse.java
package org.actics.customer.profiling.model;

public class CustomerProfileResponse {

  private String id;
  private String firstName;
  private String lastName;

  // Getter und Setter
  public String getId() {
    return id;
  }

  public void setId(String id) {
    this.id = id;
  }

  public String getFirstName() {
    return firstName;
  }

  public void setFirstName(String firstName) {
    this.firstName = firstName;
  }

  public String getLastName() {
    return lastName;
  }

  public void setLastName(String lastName) {
    this.lastName = lastName;
  }
}