# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: "A REST API for managing customer profiles compliant with FINMA's FIDLEG\
    \ requirements, including economic circumstances, risk tolerance, and more."
  title: FIDLEG Compliant Customer Profile API
  version: 1.0.0
servers:
- description: Production server
  url: https://api.example.com/v1
- description: Sandbox server
  url: https://sandbox.api.example.com/v1
paths:
  /customers:
    get:
      description: Get the details of all customer profiles without filters.
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CustomerProfileResponse'
                type: array
          description: List of customer profiles
      summary: Retrieve all customer profiles
      x-accepts:
      - application/json
    post:
      description: Create a new customer profile with details required for FIDLEG
        compliance.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerProfile'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProfileResponse'
          description: Customer profile created successfully
        "400":
          description: Invalid input data
      summary: Create a new customer profile
      x-content-type: application/json
      x-accepts:
      - application/json
  /customers/search:
    get:
      description: "Search customer profiles by attributes such as firstName, lastName,\
        \ birthdate, riskTolerance, email, or phone."
      parameters:
      - description: Filter by first name
        explode: true
        in: query
        name: firstName
        required: false
        schema:
          type: string
        style: form
      - description: Filter by last name
        explode: true
        in: query
        name: lastName
        required: false
        schema:
          type: string
        style: form
      - description: Filter by birthdate
        explode: true
        in: query
        name: birthdate
        required: false
        schema:
          format: date
          type: string
        style: form
      - description: Filter by risk tolerance
        explode: true
        in: query
        name: riskTolerance
        required: false
        schema:
          enum:
          - Low
          - Medium
          - High
          type: string
        style: form
      - description: Filter by email
        explode: true
        in: query
        name: email
        required: false
        schema:
          format: email
          type: string
        style: form
      - description: Filter by phone number
        explode: true
        in: query
        name: phone
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CustomerProfileResponse'
                type: array
          description: List of matching customer profiles
        "400":
          description: Invalid query parameters
      summary: Search for customer profiles
      x-accepts:
      - application/json
  /customers/{customerId}:
    get:
      description: Get the details of a customer profile by ID.
      parameters:
      - description: The ID of the customer
        explode: false
        in: path
        name: customerId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProfileResponse'
          description: Customer profile details
        "404":
          description: Customer not found
      summary: Retrieve a customer profile
      x-accepts:
      - application/json
components:
  schemas:
    CustomerProfile:
      example:
        firstName: John
        lastName: Doe
        birthdate: 1980-01-01
        economicCircumstances:
          annualIncome: 100000
          netWorth: 500000
          liabilities: 100000
        investmentExperience: Moderate
        riskTolerance: Medium
        investmentObjectives: Wealth accumulation
        contactDetails:
          phone: "+41791234567"
          email: john.doe@example.com
      properties:
        firstName:
          example: John
          type: string
        lastName:
          example: Doe
          type: string
        birthdate:
          example: 1980-01-01
          format: date
          type: string
        contactDetails:
          $ref: '#/components/schemas/CustomerProfile_contactDetails'
        economicCircumstances:
          $ref: '#/components/schemas/CustomerProfile_economicCircumstances'
        riskTolerance:
          enum:
          - Low
          - Medium
          - High
          example: Medium
          type: string
        investmentExperience:
          example: Moderate
          type: string
        investmentObjectives:
          example: Wealth accumulation
          type: string
      required:
      - birthdate
      - economicCircumstances
      - firstName
      - lastName
      - riskTolerance
      type: object
    CustomerProfileResponse:
      example:
        profile:
          firstName: John
          lastName: Doe
          birthdate: 1980-01-01
          economicCircumstances:
            annualIncome: 100000
            netWorth: 500000
            liabilities: 100000
          investmentExperience: Moderate
          riskTolerance: Medium
          investmentObjectives: Wealth accumulation
          contactDetails:
            phone: "+41791234567"
            email: john.doe@example.com
        id: c123456
      properties:
        id:
          example: c123456
          type: string
        profile:
          $ref: '#/components/schemas/CustomerProfile'
      type: object
    CustomerProfile_contactDetails:
      example:
        phone: "+41791234567"
        email: john.doe@example.com
      properties:
        email:
          example: john.doe@example.com
          format: email
          type: string
        phone:
          example: "+41791234567"
          type: string
      type: object
    CustomerProfile_economicCircumstances:
      example:
        annualIncome: 100000
        netWorth: 500000
        liabilities: 100000
      properties:
        annualIncome:
          example: 100000
          type: number
        netWorth:
          example: 500000
          type: number
        liabilities:
          example: 100000
          type: number
      type: object